Refactor the `db` package
