+++ db/db_ops.go
@@ -1,8 +1,6 @@
 package db
 
 import (
-	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
 
@@ -10,7 +8,6 @@ import (
 	"github.com/rs/zerolog/log"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 )
 
 var (
@@ -101,148 +98,4 @@ func CloseDB() error {
 		return err
 	}
 	return sqlDB.Close()
-}
-
-// Game represents a game record in the catalogue.
-type Game struct {
-	ID    int    `gorm:"primaryKey" json:"id"`
-	Title string `gorm:"index" json:"title"` // Indexed for faster queries
-	Data  string `json:"data"`
-}
-
-// PutInGame inserts or updates a game record in the catalogue.
-func PutInGame(id int, title, data string) error {
-	game := Game{
-		ID:    id,
-		Title: title,
-		Data:  data,
-	}
-	return upsertGame(game)
-}
-
-// upsertGame performs an upsert operation on the game record.
-func upsertGame(game Game) error {
-	if err := Db.Clauses(
-		clause.OnConflict{UpdateAll: true},
-	).Create(&game).Error; err != nil {
-		log.Error().Err(err).Msgf("Failed to upsert game with ID %d", game.ID)
-		return err
-	}
-
-	log.Info().Msgf("Game upserted successfully: ID=%d, Title=%s", game.ID, game.Title)
-	return nil
-}
-
-// EmptyCatalogue removes all records from the game catalogue.
-func EmptyCatalogue() error {
-	if err := Db.Unscoped().Where("1 = 1").Delete(&Game{}).Error; err != nil {
-		log.Error().Err(err).Msg("Failed to empty game catalogue")
-		return err
-	}
-
-	log.Info().Msg("Game catalogue emptied successfully")
-	return nil
-}
-
-// GetCatalogue retrieves all games in the catalogue.
-func GetCatalogue() ([]Game, error) {
-	var games []Game
-	if err := Db.Find(&games).Error; err != nil {
-		log.Error().Err(err).Msg("Failed to fetch games from the database")
-		return nil, err
-	}
-
-	log.Info().Msgf("Retrieved %d games from the catalogue", len(games))
-	return games, nil
-}
-
-// GetGameByID retrieves a game from the catalogue by its ID.
-func GetGameByID(id int) (*Game, error) {
-	if Db == nil {
-		return nil, fmt.Errorf("database connection is not initialized")
-	}
-
-	var game Game
-	if err := Db.First(&game, "id = ?", id).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, nil // Game not found
-		}
-		return nil, fmt.Errorf("failed to retrieve game with ID %d: %w", id, err)
-	}
-
-	return &game, nil
-}
-
-// SearchGamesByName searches for games in the catalogue by name.
-func SearchGamesByName(name string) ([]Game, error) {
-	if Db == nil {
-		return nil, fmt.Errorf("database connection is not initialized")
-	}
-
-	var games []Game
-	if err := Db.Where("title LIKE ?", "%"+name+"%").Find(&games).Error; err != nil {
-		log.Error().Err(err).Msgf("Failed to search games by name: %s", name)
-		return nil, err
-	}
-
-	return games, nil
-}
-
-// Token represents the user's authentication token data.
-type Token struct {
-	AccessToken  string `json:"access_token,omitempty"`
-	RefreshToken string `json:"refresh_token,omitempty"`
-	ExpiresAt    string `json:"expires_at,omitempty"`
-}
-
-// GetTokenRecord retrieves the token record from the database.
-func GetTokenRecord() (*Token, error) {
-	if Db == nil {
-		return nil, fmt.Errorf("database connection is not initialized")
-	}
-
-	var token Token
-	if err := Db.First(&token).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return nil, nil // Token not found
-		}
-		log.Error().Err(err).Msg("Failed to retrieve token data")
-		return nil, err
-	}
-
-	return &token, nil
-}
-
-// UpsertTokenRecord inserts or updates the token record in the database.
-func UpsertTokenRecord(token *Token) error {
-	if Db == nil {
-		return fmt.Errorf("database connection is not initialized")
-	}
-
-	var existingToken Token
-	err := Db.First(&existingToken).Error
-	if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
-		log.Error().Err(err).Msg("Failed to check existing token")
-		return err
-	}
-
-	if errors.Is(err, gorm.ErrRecordNotFound) {
-		if err := Db.Create(token).Error; err != nil {
-			log.Error().Err(err).Msg("Failed to insert new token")
-			return err
-		}
-		log.Info().Msg("Token inserted successfully")
-	} else {
-		if err := Db.Model(&existingToken).Where("1 = 1").Updates(Token{
-			AccessToken:  token.AccessToken,
-			RefreshToken: token.RefreshToken,
-			ExpiresAt:    token.ExpiresAt,
-		}).Error; err != nil {
-			log.Error().Err(err).Msg("Failed to update token")
-			return err
-		}
-		log.Info().Msg("Token updated successfully")
-	}
-
-	return nil
 }
